deleted reponse_test.py since it was testing dictionaries, but now we have objects from the response separated Initialise object from __init__.py to fix setup.py changed affected tests
