+++ tests/exceptions_test.py
@@ -3,12 +3,11 @@
 import unittest
 
 from dota2api.src import exceptions
-
+from dota2api.api import *
 
 class APIErrorTest(unittest.TestCase):
     def setUp(self):
-        import dota2api
-        self.api = dota2api.Initialise()
+        self.api = Initialise()
 
     def wrong_account_id(self):
         self.assertRaises(exceptions.APIError(), self.api.get_match_history(account_id=1))
@@ -16,8 +15,7 @@ class APIErrorTest(unittest.TestCase):
 
 class APIAuthenticationErrorTest(unittest.TestCase):
     def setUp(self):
-        import dota2api
-        self.api = dota2api.Initialise("d")
+        self.api = Initialise("d")
 
     def wrong_account_id(self):
         self.assertRaises(exceptions.APIAuthenticationError(), self.api.get_match_details())
@@ -25,8 +23,7 @@ class APIAuthenticationErrorTest(unittest.TestCase):
 
 class APITimeoutErrorTest(unittest.TestCase):
     def setUp(self):
-        import dota2api
-        self.api = dota2api.Initialise()
+        self.api = Initialise()
 
     def too_many_requests(self):
         assert True  # Can I do this?
